backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# LangGraph
langgraph>=0.2.0
langchain>=0.2.17
langchain-community>=0.2.17
tavily-python>=0.5.0

# PDF Processing
pdfplumber>=0.11.0
pymupdf>=1.24.0
pytesseract>=0.3.13
Pillow>=10.4.0
pdf2image>=1.17.0

# HTTP + Scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
playwright>=1.45.0
requests>=2.32.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.8

# Vector Store
qdrant-client>=1.11.0
sentence-transformers>=3.1.0

# ML + Explainability
xgboost>=2.1.0
lightgbm>=4.5.0
shap>=0.46.0
scikit-learn>=1.5.0
pandas>=2.2.0
openpyxl>=3.1.5
xlrd>=2.0.1
numpy>=1.26.0
networkx>=3.3.0
spacy>=3.7.5
transformers>=4.44.2

# Document Generation
python-docx>=1.1.0
reportlab>=4.2.0

# Google Gemini
google-generativeai>=0.8.0

# OpenAI SDK — used for BOTH Cerebras and GitHub Models via custom base_url
openai>=1.30.0
huggingface-hub>=0.33.2

# Utilities
python-multipart>=0.0.9
aiofiles>=24.1.0
structlog>=24.4.0

# Orchestration / Queue / Big data
celery>=5.4.0

prefect>=2.20.3
pyspark>=3.5.2
delta-spark>=3.2.0
databricks-connect>=14.3.0

# Crawl4AI + research synthesis
crawl4ai>=0.4.0
firecrawl-py>=4.18.0
anthropic>=0.28.0
tenacity>=8.3.0
